% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_circle.R
\name{circle_circle}
\alias{circle_circle}
\title{calculate circle_circle layout manually}
\usage{
circle_circle(layout, index = NULL)
}
\arguments{
\item{layout}{a layout object}

\item{index}{grouping}
}
\value{

}
\description{
calculate circle_circle layout manually
}
\examples{
  libray(tidygraph)
  n <- 1000
  microbiome <- data.frame(
    otu = paste("OTU",1:n,sep="_"),
    phylum = sample(paste("phylum",1:5,sep="_"),n,replace = T),
    class = sample(paste("class",6:30,sep="_"),n,replace=T),
    order = sample(paste("order",31:80,sep="_"),n,replace = T),
    value = runif(n,min=1,max=1000)
  )
  index_micro <- c("phylum","class","order")
  nodes_micro <- gather_graph_node(microbiome,index=index_micro, root="Bac")
  edges_micro <- gather_graph_edge(microbiome,index=index_micro, root="Bac")
  graph_micro <- tbl_graph(nodes_micro,edges_micro)
  layout_micro <- create_layout(graph_micro,layout = "circle")
  layout <- circle_circle(layout_micro,index=index_micro)
}
