% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_graph_edge.R
\name{gather_graph}
\alias{gather_graph}
\alias{gather_graph_edge}
\title{Gather graph node from data.frame}
\usage{
gather_graph_edge(df, index = NULL, root = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{index}{grouping}

\item{root}{root name}
}
\value{
a tibble
}
\description{
Gather graph node from data.frame
}
\examples{
 ibrary(ccgraph)
 data(OTU)
 edges <- gather_graph_edge(OTU,index = c("p","c","o"))
}
