% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_graph_node.R
\name{gather_graph_node}
\alias{gather_graph_node}
\title{Gather graph node from data.frame}
\usage{
gather_graph_node(df, index = NULL, value = tail(colnames(df), 1), root = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{index}{grouping}

\item{value}{}

\item{root}{root name}
}
\value{
a tibble
}
\description{
Gather graph node from data.frame
}
\examples{
 library(ccgraph)
 data(OTU)
 nodes <- gather_graph_node(OTU,index = c("p","c","o"))
}
